datasets/data_filtering/laion_aesthetic_filter/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
webdataset>=1.0.0
open-clip-torch>=2.0.0
pandas>=2.0.0
pyarrow>=12.0.0
numpy>=1.24.0
Pillow>=9.5.0
tqdm>=4.65.0
pyyaml>=6.0
matplotlib>=3.7.0
ftfy>=6.1.0
regex